+++ src/api/caidapp/tasks.py
@@ -114,6 +114,9 @@ def predict_species_on_success(
         uploaded_archive.save()
 
 
+
+
+
 def _prepare_dataframe_for_identification(mediafiles) -> dict:
     media_root = Path(settings.MEDIA_ROOT)
     csv_len = len(mediafiles)